+++ app/src/main/java/com/forrestguice/suntimeswidget/SuntimesSettingsActivity.java
@@ -272,6 +272,9 @@ public class SuntimesSettingsActivity extends PreferenceActivity implements Shar
         SuntimesWidget0.triggerWidgetUpdate(this, SuntimesWidget0_2x1.class);
         SuntimesWidget0.triggerWidgetUpdate(this, SuntimesWidget1.class);
         SuntimesWidget0.triggerWidgetUpdate(this, SuntimesWidget2.class);
+        MoonWidget0.triggerWidgetUpdate(this, MoonWidget0.class);
+        MoonWidget0.triggerWidgetUpdate(this, MoonWidget0_2x1.class);
+        MoonWidget0.triggerWidgetUpdate(this, MoonWidget0_3x1.class);
     }
 
     protected void rebuildActivity()