+++ app/src/main/java/com/forrestguice/suntimeswidget/calculator/SuntimesRiseSetDataset.java
@@ -20,14 +20,11 @@ package com.forrestguice.suntimeswidget.calculator;
 
 import android.appwidget.AppWidgetManager;
 import android.content.Context;
-import android.util.Log;
 
 import com.forrestguice.suntimeswidget.settings.WidgetSettings;
 
-import java.util.ArrayList;
 import java.util.Calendar;
 import java.util.Date;
-import java.util.GregorianCalendar;
 import java.util.TimeZone;
 
 public class SuntimesRiseSetDataset