+++ app/src/main/java/com/forrestguice/suntimeswidget/layouts/MoonLayout_1x1_0.java
@@ -1,0 +1,80 @@
+/**
+   Copyright (C) 2018 Forrest Guice
+   This file is part of SuntimesWidget.
+
+   SuntimesWidget is free software: you can redistribute it and/or modify
+   it under the terms of the GNU General Public License as published by
+   the Free Software Foundation, either version 3 of the License, or
+   (at your option) any later version.
+
+   SuntimesWidget is distributed in the hope that it will be useful,
+   but WITHOUT ANY WARRANTY; without even the implied warranty of
+   MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+   GNU General Public License for more details.
+
+   You should have received a copy of the GNU General Public License
+   along with SuntimesWidget.  If not, see <http://www.gnu.org/licenses/>.
+*/
+
+package com.forrestguice.suntimeswidget.layouts;
+
+import android.content.Context;
+import android.widget.RemoteViews;
+
+import com.forrestguice.suntimeswidget.R;
+import com.forrestguice.suntimeswidget.SuntimesUtils;
+import com.forrestguice.suntimeswidget.calculator.SuntimesMoonData;
+import com.forrestguice.suntimeswidget.settings.WidgetSettings;
+import com.forrestguice.suntimeswidget.themes.SuntimesTheme;
+
+/**
+ * Moonrise / Moonset (1x1)
+ */
+public class MoonLayout_1x1_0 extends MoonLayout
+{
+    public MoonLayout_1x1_0()
+    {
+        super();
+    }
+
+    public MoonLayout_1x1_0(int layoutID)
+    {
+        this.layoutID = layoutID;
+    }
+
+    @Override
+    public void initLayoutID()
+    {
+        this.layoutID = R.layout.layout_widget_moon_1x1_0;
+    }
+
+    @Override
+    public void updateViews(Context context, int appWidgetId, RemoteViews views, SuntimesMoonData data)
+    {
+        super.updateViews(context, appWidgetId, views, data);
+        boolean showSeconds = WidgetSettings.loadShowSecondsPref(context, appWidgetId);
+
+        SuntimesUtils.TimeDisplayText riseString = utils.calendarTimeShortDisplayString(context, data.moonriseCalendarToday(), showSeconds);
+        views.setTextViewText(R.id.text_time_moonrise, riseString.getValue());
+        views.setTextViewText(R.id.text_time_moonrise_suffix, riseString.getSuffix());
+
+        SuntimesUtils.TimeDisplayText setString = utils.calendarTimeShortDisplayString(context, data.moonsetCalendarToday(), showSeconds);
+        views.setTextViewText(R.id.text_time_moonset, setString.getValue());
+        views.setTextViewText(R.id.text_time_moonset_suffix, setString.getSuffix());
+    }
+
+    @Override
+    public void themeViews(Context context, RemoteViews views, SuntimesTheme theme)
+    {
+        super.themeViews(context, views, theme);
+        themeViewsMoonRiseSetText(context, views, theme);
+        themeViewsMoonRiseSetIcons(context, views, theme);
+    }
+
+    @Override
+    public void prepareForUpdate(SuntimesMoonData data)
+    {
+        this.layoutID = chooseMoonLayout(R.layout.layout_widget_moon_1x1_0, R.layout.layout_widget_moon_1x1_01, data);
+    }
+
+}